Split automation integration test suites
