+++ Tests/QuillCodeAppTests/WorkspaceAutomationRunIntegrationTests.swift
@@ -1,0 +1,256 @@
+import Foundation
+import XCTest
+import QuillCodeCore
+@testable import QuillCodeApp
+
+@MainActor
+final class WorkspaceAutomationRunIntegrationTests: XCTestCase {
+    func testAutomationRunCreatesFollowUpThreadAndPersistsRunMetadata() throws {
+        let source = ChatThread(title: "Launch plan", messages: [
+            .init(role: .user, content: "latest question"),
+            .init(role: .tool, content: #"{"internal":"skip"}"#),
+            .init(role: .assistant, content: "latest answer")
+        ])
+        let automation = QuillAutomation(
+            title: "Follow up: Launch plan",
+            detail: "Resume this thread with the same project, model, and context.",
+            kind: .threadFollowUp,
+            scheduleKind: .heartbeat,
+            scheduleDescription: "Manual follow-up",
+            threadID: source.id,
+            nextRunAt: Date(timeIntervalSince1970: 1)
+        )
+        let workspace = try makeAutomationWorkspace(rootState: QuillCodeRootState(
+            threads: [source],
+            selectedThreadID: source.id
+        ))
+        workspace.model.setAutomations([automation])
+        workspace.model.startSidebarSelection(selecting: source.id)
+        XCTAssertEqual(workspace.model.selectedSidebarThreadIDs(), [source.id])
+
+        XCTAssertTrue(workspace.model.runWorkspaceCommand("automation-run:\(automation.id.uuidString)", workspaceRoot: workspace.root))
+
+        let followUpID = try XCTUnwrap(workspace.model.root.selectedThreadID)
+        XCTAssertNotEqual(followUpID, source.id)
+        let followUp = try XCTUnwrap(workspace.model.root.threads.first { $0.id == followUpID })
+        XCTAssertEqual(followUp.title, "Follow-up: Launch plan")
+        XCTAssertEqual(followUp.messages.map(\.content), ["latest question", "latest answer"])
+        XCTAssertFalse(followUp.messages.contains { $0.role == .tool })
+        XCTAssertEqual(followUp.events.first?.summary, "Automation ran: Follow up: Launch plan")
+        XCTAssertEqual(followUp.events.first?.payloadJSON, automation.id.uuidString)
+
+        let savedThread = try workspace.threadStore.load(followUpID)
+        XCTAssertEqual(savedThread.title, "Follow-up: Launch plan")
+        let savedAutomation = try XCTUnwrap(try workspace.automationStore.load().first)
+        XCTAssertNotNil(savedAutomation.lastRunAt)
+        XCTAssertNil(savedAutomation.nextRunAt)
+        XCTAssertEqual(workspace.model.selectedSidebarThreadIDs(), [])
+        XCTAssertEqual(workspace.model.surface().automations.workflows.first?.statusLabel, "Ran")
+        XCTAssertEqual(workspace.model.surface().automations.statusLabel, "1 active")
+    }
+
+    func testRunWorkspaceScheduleCreatesScheduledProjectThread() throws {
+        let workspace = try makeProjectAutomationWorkspace()
+        let project = try XCTUnwrap(workspace.model.selectedProject)
+        let automation = QuillAutomation(
+            title: "Workspace check: QuillCode",
+            detail: "Create a scheduled workspace-check thread for QuillCode.",
+            kind: .workspaceSchedule,
+            scheduleKind: .cron,
+            scheduleDescription: "Manual workspace check",
+            projectID: project.id
+        )
+        workspace.model.setAutomations([automation])
+
+        XCTAssertTrue(workspace.model.runWorkspaceCommand("automation-run:\(automation.id.uuidString)", workspaceRoot: workspace.root))
+
+        let scheduledID: UUID = try XCTUnwrap(workspace.model.root.selectedThreadID)
+        let scheduled: ChatThread = try XCTUnwrap(workspace.model.root.threads.first { $0.id == scheduledID })
+        XCTAssertEqual(scheduled.title, "Scheduled check: QuillCode")
+        XCTAssertEqual(scheduled.projectID, project.id)
+        XCTAssertEqual(scheduled.messages.map(\.content), [
+            "Run the scheduled workspace check for QuillCode. Start with project status, recent changes, local actions, and anything needing attention."
+        ])
+        XCTAssertEqual(scheduled.events.first?.summary, "Automation ran: Workspace check: QuillCode")
+        XCTAssertEqual(scheduled.events.first?.payloadJSON, automation.id.uuidString)
+
+        let savedThread = try workspace.threadStore.load(scheduledID)
+        XCTAssertEqual(savedThread.title, "Scheduled check: QuillCode")
+        let savedAutomation = try XCTUnwrap(try workspace.automationStore.load().first)
+        XCTAssertNotNil(savedAutomation.lastRunAt)
+        XCTAssertNil(savedAutomation.nextRunAt)
+        XCTAssertEqual(workspace.model.surface().automations.workflows.first?.statusLabel, "Ran")
+        XCTAssertEqual(workspace.model.surface().automations.statusLabel, "1 active")
+    }
+
+    func testRunDueAutomationsRunsActiveDueThreadAndWorkspaceSchedules() throws {
+        let now = Date(timeIntervalSince1970: 100)
+        let project = ProjectRef(name: "QuillCode", path: "/tmp/QuillCode")
+        let source = ChatThread(title: "Due follow-up", messages: [
+            .init(role: .user, content: "summarize tomorrow"),
+            .init(role: .assistant, content: "I will follow up.")
+        ])
+        let due = threadFollowUpAutomation(
+            title: "Due follow-up",
+            threadID: source.id,
+            nextRunAt: Date(timeIntervalSince1970: 90)
+        )
+        let workspaceDue = QuillAutomation(
+            title: "Due workspace check",
+            detail: "Check workspace.",
+            kind: .workspaceSchedule,
+            scheduleKind: .cron,
+            scheduleDescription: "Now",
+            projectID: project.id,
+            nextRunAt: Date(timeIntervalSince1970: 85)
+        )
+        let future = threadFollowUpAutomation(
+            title: "Future follow-up",
+            detail: "Resume later.",
+            threadID: source.id,
+            scheduleDescription: "Later",
+            nextRunAt: Date(timeIntervalSince1970: 120)
+        )
+        let paused = threadFollowUpAutomation(
+            title: "Paused follow-up",
+            detail: "Do not run.",
+            status: .paused,
+            threadID: source.id,
+            nextRunAt: Date(timeIntervalSince1970: 80)
+        )
+        let unsupported = QuillAutomation(
+            title: "Due monitor",
+            detail: "Monitor runner pending.",
+            kind: .monitor,
+            scheduleKind: .event,
+            scheduleDescription: "Now",
+            nextRunAt: Date(timeIntervalSince1970: 70)
+        )
+        let workspace = try makeAutomationWorkspace(rootState: QuillCodeRootState(
+            projects: [project],
+            selectedProjectID: project.id,
+            threads: [source],
+            selectedThreadID: source.id
+        ))
+        workspace.model.setAutomations([future, paused, unsupported, due, workspaceDue])
+
+        let followUpIDs = workspace.model.runDueAutomations(now: now)
+
+        XCTAssertEqual(followUpIDs.count, 2)
+        let followUp = try XCTUnwrap(workspace.model.root.threads.first { $0.title == "Follow-up: Due follow-up" })
+        XCTAssertEqual(followUp.title, "Follow-up: Due follow-up")
+        let workspaceCheck = try XCTUnwrap(workspace.model.root.threads.first { $0.title == "Scheduled check: QuillCode" })
+        XCTAssertEqual(workspaceCheck.projectID, project.id)
+        XCTAssertTrue([followUp.id, workspaceCheck.id].contains(try XCTUnwrap(workspace.model.root.selectedThreadID)))
+
+        let savedAutomations = try workspace.automationStore.load()
+        let savedDue = try XCTUnwrap(savedAutomations.first { $0.id == due.id })
+        XCTAssertNotNil(savedDue.lastRunAt)
+        XCTAssertNil(savedDue.nextRunAt)
+        let savedWorkspaceDue = try XCTUnwrap(savedAutomations.first { $0.id == workspaceDue.id })
+        XCTAssertNotNil(savedWorkspaceDue.lastRunAt)
+        XCTAssertNil(savedWorkspaceDue.nextRunAt)
+        XCTAssertEqual(savedAutomations.first { $0.id == future.id }?.nextRunAt, future.nextRunAt)
+        XCTAssertEqual(savedAutomations.first { $0.id == paused.id }?.nextRunAt, paused.nextRunAt)
+        XCTAssertEqual(savedAutomations.first { $0.id == unsupported.id }?.nextRunAt, unsupported.nextRunAt)
+    }
+
+    func testDueRecurringWorkspaceScheduleRunsAndAdvancesNextRun() throws {
+        let runAt = Date(timeIntervalSince1970: floor(Date().timeIntervalSince1970) + 60)
+        let workspace = try makeProjectAutomationWorkspace()
+        let project = try XCTUnwrap(workspace.model.selectedProject)
+        let recurrence = QuillAutomationRecurrence(interval: 2, unit: .hours)
+        let recurring = QuillAutomation(
+            title: "Recurring workspace check",
+            detail: "Check workspace.",
+            kind: .workspaceSchedule,
+            scheduleKind: .cron,
+            scheduleDescription: recurrence.scheduleDescription,
+            projectID: project.id,
+            nextRunAt: runAt.addingTimeInterval(-10),
+            recurrence: recurrence
+        )
+        workspace.model.setAutomations([recurring])
+
+        let reports = workspace.model.runDueAutomationReports(now: runAt)
+
+        XCTAssertEqual(reports.count, 1)
+        XCTAssertEqual(workspace.model.root.threads.filter { $0.title == "Scheduled check: QuillCode" }.count, 1)
+        let saved = try XCTUnwrap(try workspace.automationStore.load().first)
+        XCTAssertNotNil(saved.lastRunAt)
+        XCTAssertEqual(
+            try XCTUnwrap(saved.nextRunAt).timeIntervalSince1970,
+            recurrence.nextRun(after: runAt).timeIntervalSince1970,
+            accuracy: 0.001
+        )
+        XCTAssertEqual(saved.recurrence, recurrence)
+        XCTAssertEqual(workspace.model.surface().automations.workflows.first?.statusLabel, "Active")
+
+        _ = workspace.model.runDueAutomations(now: runAt.addingTimeInterval(60))
+
+        XCTAssertEqual(workspace.model.root.threads.filter { $0.title == "Scheduled check: QuillCode" }.count, 1)
+    }
+
+    func testRunDueAutomationReportsDescribeCreatedFollowUps() throws {
+        let now = Date(timeIntervalSince1970: 100)
+        let source = ChatThread(title: "Due follow-up", messages: [
+            .init(role: .user, content: "summarize tomorrow"),
+            .init(role: .assistant, content: "I will follow up.")
+        ])
+        let due = threadFollowUpAutomation(
+            title: "Due follow-up",
+            threadID: source.id,
+            nextRunAt: Date(timeIntervalSince1970: 90)
+        )
+        let workspace = try makeAutomationWorkspace(rootState: QuillCodeRootState(
+            threads: [source],
+            selectedThreadID: source.id
+        ))
+        workspace.model.setAutomations([due])
+
+        let reports = workspace.model.runDueAutomationReports(now: now)
+
+        let report = try XCTUnwrap(reports.first)
+        XCTAssertEqual(reports.count, 1)
+        XCTAssertEqual(report.automationID, due.id)
+        XCTAssertEqual(report.title, "QuillCode follow-up ready")
+        XCTAssertEqual(report.body, "Follow-up: Due follow-up was created from Due follow-up.")
+        XCTAssertEqual(workspace.model.root.selectedThreadID, report.followUpThreadID)
+        XCTAssertEqual(workspace.model.root.threads.first?.id, report.followUpThreadID)
+        XCTAssertEqual(workspace.model.root.threads.first?.title, "Follow-up: Due follow-up")
+
+        let savedDue = try XCTUnwrap(try workspace.automationStore.load().first { $0.id == due.id })
+        XCTAssertNotNil(savedDue.lastRunAt)
+        XCTAssertNil(savedDue.nextRunAt)
+    }
+
+    func testRunDueAutomationsHonorsLimit() throws {
+        let now = Date(timeIntervalSince1970: 100)
+        let source = ChatThread(title: "Launch plan", messages: [
+            .init(role: .user, content: "follow up"),
+            .init(role: .assistant, content: "Noted.")
+        ])
+        let first = threadFollowUpAutomation(
+            title: "First follow-up",
+            threadID: source.id,
+            nextRunAt: Date(timeIntervalSince1970: 80)
+        )
+        let second = threadFollowUpAutomation(
+            title: "Second follow-up",
+            threadID: source.id,
+            nextRunAt: Date(timeIntervalSince1970: 90)
+        )
+        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
+            threads: [source],
+            selectedThreadID: source.id
+        ))
+        model.setAutomations([second, first])
+
+        let followUpIDs = model.runDueAutomations(now: now, limit: 1)
+
+        XCTAssertEqual(followUpIDs.count, 1)
+        XCTAssertEqual(model.automations.items.first { $0.id == first.id }?.lastRunAt != nil, true)
+        XCTAssertNil(model.automations.items.first { $0.id == first.id }?.nextRunAt)
+        XCTAssertEqual(model.automations.items.first { $0.id == second.id }?.nextRunAt, second.nextRunAt)
+    }
+}
