+++ Tests/QuillCodeParityTests/ParityWorkspaceIntegrationGateTests.swift
@@ -113,12 +113,21 @@ final class ParityWorkspaceIntegrationGateTests: QuillCodeParityTestCase {
 
     func testWorkspaceAutomationIntegrationTestsOwnModelAutomationFlows() throws {
         let modelTests = try Self.appTestSourceText(named: "WorkspaceModelTests.swift")
-        let automationIntegrationTests = try Self.appTestSourceText(named: "WorkspaceAutomationIntegrationTests.swift")
+        let automationCommandTests = try Self.appTestSourceText(named: "WorkspaceAutomationIntegrationTests.swift")
+        let automationSchedulingTests = try Self.appTestSourceText(named: "WorkspaceAutomationSchedulingIntegrationTests.swift")
+        let automationRunTests = try Self.appTestSourceText(named: "WorkspaceAutomationRunIntegrationTests.swift")
+        let automationSupport = try Self.appTestSourceText(named: "WorkspaceAutomationIntegrationTestSupport.swift")
 
-        XCTAssertTrue(automationIntegrationTests.contains("testAutomationCommandsCreatePauseResumeAndDeletePersistedFollowUp"), "Automation command persistence should live in focused automation integration tests.")
-        XCTAssertTrue(automationIntegrationTests.contains("testSlashFollowUpSchedulesCurrentThread"), "Slash follow-up scheduling should live in focused automation integration tests.")
-        XCTAssertTrue(automationIntegrationTests.contains("testRunDueAutomationsRunsActiveDueThreadAndWorkspaceSchedules"), "Due automation runs should live in focused automation integration tests.")
-        XCTAssertTrue(automationIntegrationTests.contains("testRunDueAutomationsHonorsLimit"), "Due automation limit integration should live in focused automation integration tests.")
+        XCTAssertTrue(automationCommandTests.contains("testAutomationCommandsCreatePauseResumeAndDeletePersistedFollowUp"), "Automation command persistence should live in focused automation command integration tests.")
+        XCTAssertTrue(automationCommandTests.contains("testCreateWorkspaceScheduleCommandPersistsSelectedProjectAutomation"), "Manual workspace schedule creation should live with automation command persistence.")
+        XCTAssertFalse(automationCommandTests.contains("testSlashFollowUpSchedulesCurrentThread"), "Scheduling flows should not live in the command/persistence automation suite.")
+        XCTAssertFalse(automationCommandTests.contains("testRunDueAutomationsRunsActiveDueThreadAndWorkspaceSchedules"), "Due-run flows should not live in the command/persistence automation suite.")
+        XCTAssertTrue(automationSchedulingTests.contains("testSlashFollowUpSchedulesCurrentThread"), "Slash follow-up scheduling should live in focused automation scheduling tests.")
+        XCTAssertTrue(automationSchedulingTests.contains("testNaturalLanguageRecurringWorkspaceChecksPersistRecurrence"), "Recurring schedule parsing should live in focused automation scheduling tests.")
+        XCTAssertTrue(automationRunTests.contains("testRunDueAutomationsRunsActiveDueThreadAndWorkspaceSchedules"), "Due automation runs should live in focused automation run tests.")
+        XCTAssertTrue(automationRunTests.contains("testRunDueAutomationsHonorsLimit"), "Due automation limit integration should live in focused automation run tests.")
+        XCTAssertTrue(automationSupport.contains("func makeAutomationWorkspace"), "Automation integration fixtures should be shared, not copied between focused suites.")
+        XCTAssertTrue(automationSupport.contains("func threadFollowUpAutomation"), "Thread follow-up fixtures should be shared by scheduling and run tests.")
         XCTAssertFalse(modelTests.contains("testAutomationCommandsCreatePauseResumeAndDeletePersistedFollowUp"), "WorkspaceModelTests should not own automation command persistence flows.")
         XCTAssertFalse(modelTests.contains("testSlashFollowUpSchedulesCurrentThread"), "WorkspaceModelTests should not own slash follow-up scheduling flows.")
         XCTAssertFalse(modelTests.contains("testRunDueAutomationsRunsActiveDueThreadAndWorkspaceSchedules"), "WorkspaceModelTests should not own due automation run flows.")
