+++ docs/DECISIONS.md
@@ -227,7 +227,7 @@
 - Workspace project extension integration coverage lives in `WorkspaceProjectExtensionIntegrationTests`, not `WorkspaceModelTests`. Extension manifest loading into project/secondary-pane surfaces, update-command execution, metadata refresh, and success/failure transcript notices cross the workspace model and project metadata loader; keep those flows together outside the model monolith while low-level manifest parsing and shell-call shape tests stay focused.
 - Workspace slash-command integration coverage lives in `WorkspaceSlashCommandIntegrationTests`, not `WorkspaceModelTests`. Command-palette slash prefills, core slash dispatch, local environment action slash execution, local model/mode/thread lifecycle commands, context compaction, and status transcript assertions cross composer submission, command dispatch, workspace surfaces, and transcript creation; keep those flows together while pure transcript copy stays in `WorkspaceSlashCommandTranscriptPlannerTests` and remote/automation-specific slash flows stay with their broader owning features until extracted.
 - Workspace local environment integration coverage lives in `WorkspaceLocalEnvironmentIntegrationTests`, not `WorkspaceModelTests`. Local environment action loading into project state, command-palette execution, metadata-backed environment redaction, bounded working directories, and timeouts cross the workspace model, metadata loader, shell-call planner, and tool cards; keep those flows together outside the model monolith while direct loader behavior stays focused. `/env` transcript behavior belongs to `WorkspaceSlashCommandIntegrationTests` with the rest of slash-command routing.
-- Workspace automation integration coverage lives in `WorkspaceAutomationIntegrationTests`, not `WorkspaceModelTests`. Automation persistence, create/pause/resume/delete commands, scheduled and natural-language follow-ups/workspace checks, slash scheduling commands, due automation execution, recurrence advancement, and run reports cross workspace model, stores, surfaces, transcript creation, and thread persistence; keep those flows together while pure automation record and run-draft logic stays in `WorkspaceAutomationEngineTests`.
+- Workspace automation integration coverage is split by workflow family, not collected in `WorkspaceModelTests` or one oversized automation suite. `WorkspaceAutomationIntegrationTests` owns command/persistence flows, `WorkspaceAutomationSchedulingIntegrationTests` owns concrete, natural-language, recurring, and slash scheduling flows, and `WorkspaceAutomationRunIntegrationTests` owns manual and due-run execution, recurrence advancement, reports, and limits. Shared fixtures live in `WorkspaceAutomationIntegrationTestSupport`, while pure automation record and run-draft logic stays in `WorkspaceAutomationEngineTests`.
 - Workspace terminal integration coverage lives in `WorkspaceTerminalIntegrationTests`, not `WorkspaceModelTests`. Local terminal execution, SSH Remote terminal execution, streaming output, cwd/environment persistence, clear-history behavior, selected-project resets, cancellation, and stop-all behavior cross the workspace model, terminal engine, shell/SSH runners, terminal surface, and async task lifecycle; keep those flows together outside the model monolith while pure terminal state transitions stay in `WorkspaceTerminalEngineTests`.
 - Shared workspace integration fixtures live in `WorkspaceModelIntegrationTestSupport` only while they are genuinely cross-domain. Fake SSH, fake GitHub CLI, temporary git repositories, and fixed/recording LLM clients are common enough to share, but new broad helpers should move to domain-specific support files once they become owned by one integration cluster.
 - Focused workspace unit suites should use `makeQuillCodeTestDirectory()` instead of private temporary-directory helpers. The shared helper registers teardown cleanup, keeps fixture policy in one place, and avoids small leaks when focused tests are split out of broader integration suites.
